# api-specs/cart-service-openapi.yaml
openapi: 3.0.3
info:
  title: Cart Service API
  description: |
    API for managing shopping carts in the Online Shopping Platform.
    Supports adding/removing items, applying coupons, and cart management.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: platform@shopping.com

servers:
  - url: http://localhost:8082
    description: Local development
  - url: https://api.shopping.com
    description: Production

tags:
  - name: Cart
    description: Shopping cart operations

paths:
  /api/v1/cart:
    get:
      tags: [Cart]
      summary: Get current user cart
      description: Retrieve the shopping cart for the authenticated user
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/cart/add:
    post:
      tags: [Cart]
      summary: Add item to cart
      description: Add a product to the user's shopping cart
      operationId: addToCart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '200':
          description: Item added, updated cart returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request (e.g., quantity < 1)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/cart/update:
    put:
      tags: [Cart]
      summary: Update cart item quantity
      operationId: updateCartItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemRequest'
      responses:
        '200':
          description: Cart updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Item not in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/cart/remove:
    delete:
      tags: [Cart]
      summary: Remove item from cart
      operationId: removeFromCart
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the product to remove
      responses:
        '200':
          description: Item removed, updated cart returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Item not in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/cart/apply-coupon:
    post:
      tags: [Cart]
      summary: Apply coupon to cart
      operationId: applyCoupon
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyCouponRequest'
      responses:
        '200':
          description: Coupon applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid or expired coupon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/cart/clear:
    delete:
      tags: [Cart]
      summary: Clear entire cart
      operationId: clearCart
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cart cleared
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Cart:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: number
          format: double
          example: 159.98
        discount:
          type: number
          format: double
          example: 10.00
        total:
          type: number
          format: double
          example: 149.98
        couponCode:
          type: string
          nullable: true
        itemCount:
          type: integer
          example: 3
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CartItem:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        productName:
          type: string
        imageUrl:
          type: string
          format: uri
        price:
          type: number
          format: double
        quantity:
          type: integer
          minimum: 1
        lineTotal:
          type: number
          format: double

    AddToCartRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 99
          example: 1

    UpdateCartItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 99

    ApplyCouponRequest:
      type: object
      required: [couponCode]
      properties:
        couponCode:
          type: string
          minLength: 3
          maxLength: 20
          pattern: "^[A-Z0-9-]+$"
          example: "SAVE10"

    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string

    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  rejectedValue:
                    type: string