# backend/cart-service/src/main/java/com/shopping/cart/dto/UpdateCartItemRequest.java
package com.shopping.cart.dto;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.NotNull;
import java.util.UUID;

public record UpdateCartItemRequest(
    @NotNull(message = "Product ID is required")
    UUID productId,

    @NotNull(message = "Quantity is required")
    @Min(value = 1, message = "Quantity must be at least 1")
    @Max(value = 99, message = "Quantity cannot exceed 99")
    Integer quantity
) {}